EPMRPP-113528 || Redirect user to newly created organization
