+++ app/src/controllers/organization/users/sagas.js
@@ -19,7 +19,7 @@ import { URLS } from 'common/urls';
 import { put, select, takeEvery, call, all } from 'redux-saga/effects';
 import { querySelector } from './selectors';
 import { fetchOrganizationUsersAction } from './actionCreators';
-import { withActiveOrganization } from '../sagas';
+import { withActiveOrganization } from '../withActiveOrganizationSaga';
 import { showSuccessNotification, showErrorNotification } from 'controllers/notification';
 import { fetch } from 'common/utils';
 import { userInfoSelector } from 'controllers/user';
